+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-# CHANGELOG for sbt-extras
+# CHANGELOG for sbt-extras cookbook
 
 ## 0.1.0:
 
-* Initial release of sbt-extras
+* Initial release